% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defseqtypes}
\alias{defseqtypes}
\title{Possible sequence types for a specified number of targets}
\usage{
defseqtypes(J)
}
\arguments{
\item{J}{(integer), number of targets.}
}
\value{
a matrix containing, in row, the possible sequence types and in columns which target is covered by each sequence type
}
\description{
Possible sequence types for a specified number of targets
}
