% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectOptPobs.R
\name{DirectOptPobs}
\alias{DirectOptPobs}
\title{Target amplification probabilities from ddPCR on full sequences}
\usage{
DirectOptPobs(NY, ObsTypes, lambda0 = NULL, prec = 100, evallam = F)
}
\arguments{
\item{NY}{numerical vector containing the number of each observation type.}

\item{ObsTypes}{binary matrix containing, for each observation type (in row), if each target is present (1) of absent (0).}

\item{lambda0}{(optional) numeric, the average number of sequences in one partition of ddPCR. If not provided, it will be estimated together with the amplification probabilities.}

\item{prec}{(optional) integer, the higher the closer the estimated average number of sequences in one partition of ddPCR should be from lambda0 (if provided).}

\item{evallam}{(optional) bolean, should lambda0 be used as is of should it be estimated along with amplification probabilities? Is automatically switched to TRUE if no lambda0 is provided.}
}
\value{
a vector containing the amplification probability for each target ; also returns the average number of sequences in one partition of ddPCR is evallam=TRUE or if lambda0 is not provided.
}
\description{
Direct maximization of the Likelihood and of the
expected number of sequence per partition
with regard to target amplification probabilities
in a special case when we know that all sequences cover all the targets
}
