% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptProplEM.R
\name{LatLibrary}
\alias{LatLibrary}
\title{Latente types library building}
\usage{
LatLibrary(Nit = 1000, NK = 1e+05, lambda, L)
}
\arguments{
\item{Nit}{(integer), number of sequence type proportion generation.}

\item{NK}{(integer), number of latent vector generation per generated proportion. Nk has to be a multiple of Nit.}

\item{lambda}{(numeric), average number of sequences in one partition of ddPCR.}

\item{L}{(interger) number of sequence types}
}
\value{
a matrix containing all the possibilities for the number of each sequence type inside a partition.
}
\description{
Latente types library building
}
