% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptPobsSEM.R
\name{LatLibraryPobs}
\alias{LatLibraryPobs}
\title{Latente types library building}
\usage{
LatLibraryPobs(Nit = 1000, NK = 1e+05, lambda, propl)
}
\arguments{
\item{Nit}{(integer), number of sequence type proportion generation.}

\item{NK}{(integer), number of latent vector generation per generated proportion. Nk has to be a multiple of Nit.}

\item{lambda}{(numeric) average number of sequences in one partition of ddPCR.}

\item{propl}{(vector) sequence types frequencies.}
}
\value{
a matrix containing all the possibilities for the number of each sequence type inside a partition.
}
\description{
When proportions of sequence types are unknown
}
