% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptPobsSEM.R
\name{LatLibraryPobsLam}
\alias{LatLibraryPobsLam}
\title{Latente types library building}
\usage{
LatLibraryPobsLam(Nit = 1000, NK = 1e+05, lambda0 = 0.45, prec = 100, propl)
}
\arguments{
\item{Nit}{(integer), number of sequence type proportion generation.}

\item{NK}{(integer), number of latent vector generation per generated proportion. Nk has to be a multiple of Nit.}

\item{lambda0}{(numeric), a starting point for average number of sequences in one partition of ddPCR.}

\item{prec}{(integer), the higher the closer the estimated average number of sequences in one partition of ddPCR should be from lambda0 (if provided).}

\item{propl}{(vector) sequence types frequencies.}
}
\value{
a matrix containing all the possibilities for the number of each sequence type inside a partition.
}
\description{
When proportions of sequence types and average number of sequences per partition are unknown
}
