% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptPobsSEM.R
\name{OptPobsSEM}
\alias{OptPobsSEM}
\title{SEM algorithm for amplification probability learning}
\usage{
OptPobsSEM(
  NY,
  ObsTypes,
  propl,
  coverseq,
  LatTypes = NULL,
  lambda0 = NULL,
  prec = 100,
  evallam = F,
  Nit = 30,
  plotopt = T,
  pobsinit = NULL
)
}
\arguments{
\item{NY}{(numerical vector) contains the number of each observation type.}

\item{ObsTypes}{(binary matrix) contains, for each observation type (in row), if each target is present (1) of absent (0).}

\item{propl}{(vector) sequence types frequencies.}

\item{coverseq}{(numeric matrix) targets covered by each of those sequence types.}

\item{LatTypes}{(optional) matrix, contains all the possibilities for the number of each sequence type inside a partition.}

\item{lambda0}{(optional) numeric, the average number of sequences in one partition of ddPCR. If not provided, it will be estimated together with the amplification probabilities.}

\item{prec}{(optional) integer, the higher the closer the estimated average number of sequences in one partition of ddPCR should be from lambda0 (if provided).}

\item{evallam}{(optional) bolean, should lambda0 be used as is of should it be estimated along with amplification probabilities? Is automatically switched to TRUE if no lambda0 is provided.}

\item{Nit}{(optional) integer, maximal number of iteration in the SEM algorithm}

\item{plotopt}{(optional) bolean, should a parameter convergence plot be drawn?}

\item{pobsinit}{(optional, numerical vector) initial values for target-specific amplification probabilities}
}
\value{
target-specific amplification probabilities ($pobs),
their values all along the SEM ($POBS),
if lambda is not provided, its values all along the SEM ($LAMBDA),
the complete log-likelohood all along the SEM ($LLC)
}
\description{
Learning of target-specific amplification probability when sequence types proportions are known
}
