% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptProplEM.R
\name{OptProplEM}
\alias{OptProplEM}
\title{EM algorithm estimating proportions of each sequence type}
\usage{
OptProplEM(
  NY,
  ObsTypes,
  LatTypes = NULL,
  pobs = NULL,
  lambda0 = NULL,
  prec = 100,
  propl0 = NULL,
  Nit = 300,
  plotopt = T,
  evallam = F,
  convdiag = 1e-07,
  convdiaglam = 0.001
)
}
\arguments{
\item{NY}{numerical vector, contains the number of each observation type.}

\item{ObsTypes}{binary matrix, contains, for each observation type (in row), if each target is present (1) of absent (0).}

\item{LatTypes}{(optional) matrix, contains all the possibilities for the number of each sequence type inside a partition.}

\item{pobs}{(optional) vector, amplification probability for each target}

\item{lambda0}{(optional) numeric, the average number of sequences in one partition of ddPCR. If not provided, it will be estimated together with the amplification probabilities.}

\item{prec}{(optional) integer, the higher the closer the estimated average number of sequences in one partition of ddPCR should be from lambda0 (if provided).}

\item{propl0}{(optional) vector, initial value for the proportions of each possible sequence type}

\item{Nit}{(optional) integer, maximal number of iteration in the EM algorithm}

\item{plotopt}{(optional) bolean, should a parameter convergence plot be drawn?}

\item{evallam}{(optional) bolean, should lambda0 be used as is of should it be estimated along with amplification probabilities? Is automatically switched to TRUE if no lambda0 is provided.}

\item{convdiag}{(optional) numeric, under which limit should the algorithm consider it has converged with respect to proportions learning?}

\item{convdiaglam}{(optional) numeric, under which limit should the algorithm consider it has converged with respect to lambda learning?}
}
\value{
a list of class ProplEM containing
the successive values of the learned sequence types frequencies in the EM algorithm ($PROPL),
the final learned sequence types frequencies ($propl),
the target covered by each of those sequence types ($coverseq),
the average number of sequences in a partition ($lambda),
the successive values of the average number of sequences in a partition in the EM algorithm ($LAMBDA)
}
\description{
EM algorithm estimating proportions of each sequence type
}
