% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SimulObs}
\alias{SimulObs}
\title{Simulation of observations from the model}
\usage{
SimulObs(N, lambda, propl, pobs)
}
\arguments{
\item{N}{(integer) to tal number of partitions}

\item{lambda}{(numeric) average number of sequences per partition}

\item{propl}{(numeric vector) proportion of each possible sequence type in the order automatically generated by the defseqtypes function}

\item{pobs}{(numeric vector) amplification probability of each target}
}
\value{
a vector containing the number of observation for each observation type ($NY)
along with those observation types ($ObsTypes) in terms of targets coverage
}
\description{
Simulation of observations from the model
}
